should be int!

+++ ViewRenderer.hpp
@@ -99,8 +99,8 @@ protected:
 	const Thing* getThing(int id) const { for (const Thing& t : things) if (t.type == id) return &t; return nullptr; }
 	void updatePlayerSubSectorHeight();
 	
-	bool doesLineIntersect(float x1, float y1, float x2, float y2) const;
-	void findIntersectingNodes(int n, float x1, float y1, float x2, float y2, std::vector<const Linedef*>& out) const;
+	bool doesLineIntersect(int x1, int y1, int x2, int y2) const;
+	void findIntersectingNodes(int n, int x1, int y1, int x2, int y2, std::vector<const Linedef*>& out) const;
 	bool isPointOnLeftSide(const Viewpoint &v, int node) const;
 	
 	void renderBSPNodes(int iNodeID);
